requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# Database
mysqlclient==2.2.0

# Authentication
django-allauth==0.57.0
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.0
django-otp>=1.6.3

# File Storage (Cloudflare R2)
django-storages==1.14.2
boto3==1.29.7

# Payment Gateway
razorpay==1.4.1

# Admin & UI
django-unfold==0.40.0
# django-crispy-forms must be >=2.3 for crispy-bootstrap5 2024.10
# Use a compatible release range to avoid resolution conflicts.
django-crispy-forms>=2.3,<3.0
# Use an available crispy-bootstrap5 release (2.0.0 is not published on PyPI).
# Pick a recent compatible tag. Available versions include 2023.10, 2024.2, 2024.10, 2025.4, 2025.6.
crispy-bootstrap5==2024.10

# API & CORS
django-cors-headers==4.3.1
drf-spectacular==0.27.0
django-filter==24.3

# Rate Limiting
django-ratelimit==4.1.0

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Audio Processing
mutagen==1.47.0

# Monitoring & Logging
django-extensions==3.2.3

# Production Server
gunicorn==21.2.0
uvicorn[standard]==0.24.0
whitenoise==6.6.0

# Security
django-environ==0.11.2

# Template tweaks used in templates (django-widget-tweaks provides the `widget_tweaks` tag)
django-widget-tweaks
